# tooling/sparta/packages/utils/openapi/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sparta API",
    "version": "1.0.0",
    "description": "…"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Sparta API"
    }
  ],
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Error message describing the issue."
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
...
      "HumanPassport": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Current status of verification (e.g., pending_signature, verification_complete)"
          },
          "score": {
            "type": "number",
            "nullable": true,
            "description": "Passport score if verification completed"
          },
          "lastVerificationTime": {
            "type": "number",
            "format": "int64",
            "nullable": true,
            "description": "Timestamp (ms) when verification was last completed"
          },
          "verificationId": {
            "type": "string",
            "nullable": true,
            "description": "ID used for the verification process"
          },
          "interactionToken": {
            "type": "string",
            "nullable": true,
            "description": "Discord interaction token for UI updates"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "discordUserId": {
            "type": "string",
            "description": "Discord user ID (Primary Key)"
          },
          "discordUsername": {
            "type": "string",
            "description": "Discord username"
          },
          "walletAddress": {
            "type": "string",
            "nullable": true,
            "description": "Ethereum address (verified through passport)"
          },
          "role": {
            "type": "string",
            "nullable": true,
            "description": "User role within the system"
          },
          "humanPassport": {
            "$ref": "#/components/schemas/HumanPassport",
            "nullable": true
          },
          "createdAt": {
            "type": "number",
            "format": "int64",
            "description": "Timestamp when user was created"
          },
          "updatedAt": {
            "type": "number",
            "format": "int64",
            "description": "Timestamp when user was last updated"
          }
        },
        "required": [
          "discordUserId",
          "discordUsername",
          "createdAt",
          "updatedAt"
        ]
      }
    }
  },
  "paths": {
...
    "/api/users": {
      "get": {
        "summary": "Get all users",
        "description": "Retrieve a list of all users",
        "tags": [
          "Users"
        ],
        "operationId": "getAllUsers",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Success status"
                    },
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "description": "Create a new user profile",
        "tags": [
          "Users"
        ],
        "operationId": "createUser",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "discordUserId",
                  "discordUsername"
                ],
                "properties": {
                  "discordUserId": {
                    "type": "string",
                    "description": "Discord user ID"
                  },
                  "discordUsername": {
                    "type": "string",
                    "description": "Discord username"
                  },
                  "walletAddress": {
                    "type": "string",
                    "description": "Ethereum wallet address"
                  },
                  "role": {
                    "type": "string",
                    "description": "User role"
                  },
                  "humanPassport": {
                    "type": "object",
                    "description": "Human passport verification data"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Bad request - Missing required fields or wallet already in use"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/users/discord/{discordUserId}": {
      "get": {
        "summary": "Get a specific user by Discord user ID",
        "description": "Retrieve a user by their Discord user ID",
        "tags": [
          "Users"
        ],
        "operationId": "getUserByDiscordId",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "discordUserId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Discord user ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Success status"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a user by Discord user ID",
        "description": "Update an existing user's information",
        "tags": [
          "Users"
        ],
        "operationId": "updateUserByDiscordId",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "discordUserId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Discord user ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "discordUsername": {
                    "type": "string",
                    "description": "Discord username"
                  },
                  "walletAddress": {
                    "type": "string",
                    "description": "Ethereum wallet address"
                  },
                  "role": {
                    "type": "string",
                    "description": "User role"
                  },
                  "humanPassport": {
                    "type": "object",
                    "description": "Human passport verification data"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          },
          "400": {
            "description": "Bad request or wallet already in use"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "summary": "Delete a user by Discord user ID",
        "description": "Delete a user by their Discord user ID",
        "tags": [
          "Users"
        ],
        "operationId": "deleteUserByDiscordId",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "discordUserId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Discord user ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/users/wallet/{walletAddress}": {
      "get": {
        "summary": "Get a user by wallet address",
        "description": "Retrieve a user by their Ethereum wallet address",
        "tags": [
          "Users"
        ],
        "operationId": "getUserByWalletAddress",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "walletAddress",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Ethereum wallet address"
          }
        ],
        "responses": {
          "200": {
            "description": "User profile or \"wallet not registered\" if address isn't used",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Success status"
                    },
                    "isRegistered": {
                      "type": "boolean",
                      "description": "Whether the wallet is registered"
                    },
                    "message": {
                      "type": "string",
                      "description": "Informational message"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Missing wallet address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}